requirements.txt
# crypto-ttt-regime dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
scipy>=1.11.0
tqdm>=4.65.0
pyarrow>=13.0.0
Pillow>=10.0.0
streamlit>=1.28.0
python-binance>=1.0.17
